Added HTML files with corresponding CSS styles.
Added HTML files in the "templates" folder and their corresponding CSS files in the "static/styles" folder to enhance the project's front end.

HTML & CSS files for user signup, login, editing to-do list items, the homepage of the to-do list (index.html), and a base template for all files to inherit from.