Implemented User Signup, Login, and To-Do List Functionality
Implemented user signup and login functionalities to allow users to create accounts and access the application securely.

Added the ability for users to create, delete, edit, and mark items as completed on their personal to-do lists.

Updated the URLs configuration in urls.py to correctly route requests to the new views created in views.py.

Added a date feature that displays the current date to users for reference.

Created the homepage (index.html) to serve as the landing page after users log in, displaying their personal to-do list.

Defined database models in models.py, including a custom user model, date model, to-do list model, and item model.

Created forms in forms.py to facilitate adding items to the to-do list, user sign-up, and user login functionalities.